Move unit tests to one placce
